requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
playwright>=1.40.0
apscheduler>=3.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
openai>=1.0.0
python-dotenv>=1.0.0
